se han agergado los reportes del cliente y algunos efectos xD, esta estructura es la que se aplicara a los demas, posterior a la creación de los reportes se procedera a dar los retoques finales

+++ Banco_ElBilleton/src/main/java/Modelo/Entidades/Objetos/Cuenta.java
@@ -12,11 +12,11 @@ package Modelo.Entidades.Objetos;
 public class Cuenta{
     private int numeroCuenta;
     private int codigoDueno;    
-    private int monto;
+    private double monto;
     private String fechaCreacion;
     private String estado;
     
-    public Cuenta(int elNumeroCuenta, int elCodigoDueno, int elMonto, String laFechaCreacion, String elEstado){
+    public Cuenta(int elNumeroCuenta, int elCodigoDueno, double elMonto, String laFechaCreacion, String elEstado){
         numeroCuenta = elNumeroCuenta;
         codigoDueno = elCodigoDueno;
         monto = elMonto;
@@ -36,7 +36,7 @@ public class Cuenta{
         return codigoDueno;
     }
     
-    public int getMonto(){
+    public double getMonto(){
         return monto;
     }
